Correct deployment instructions
The existing instructions assume the web app deployment is already linked for deployments through a git commit hook. However these instructions start out by having the student download a zipfile, therefore the source is not linked to a git repository and the deployment needs to be initiated manually from local source.